latest update

+++ grails-app/controllers/ProductSubCategoryController.groovy
@@ -13,19 +13,34 @@ class ProductSubCategoryController extends BaseController{
 
     def list() {
         try{
-        def productSubCategoryList=ProductSubCategory.list()
+            if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+
+                def productSubCategoryList=ProductSubCategory.list()
         render(view: "list",model: [productSubCategoryList:productSubCategoryList])
     }
+            else{
+                redirect(action: "adminLoginForm",controller: "login")
+
+            }
+        }
         catch (Exception e){
             redirect(action: "notfound",controller: "errorPage")
         }
     }
     def create(){
+        if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+        render(view: "create")
+        }
+        else{
+            redirect(action: "adminLoginForm",controller: "login")
 
+        }
     }
     def save(){
         try{
-        if(!params.id){
+            if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+
+                if(!params.id){
             def productSubCategoryInstance=new ProductSubCategory()
             productSubCategoryInstance.subCategoryName=params.subCategoryName
             productSubCategoryInstance.productSubCategorySpecify=ProductSubCategorySpecify.get(params.productSubCategorySpecify)
@@ -56,56 +71,30 @@ if(productSubCategoryInstance){
 
 }
         }}
+            else{
+                redirect(action: "adminLoginForm",controller: "login")
+
+            }
+        }
         catch (Exception e){
             redirect(action: "notfound",controller: "errorPage")
 
         }
     }
     def uploadCoverImage(){
-        def mp = (MultipartHttpServletRequest) request
-        CommonsMultipartFile file = (CommonsMultipartFile) mp.getFile("coverImageName")
-        def fileName=file.originalFilename
-        def homeDir = new File(System.getProperty("user.home"))
-        File theDir = new File(homeDir,"yarsaa");
-        if (! theDir.exists()){
-            theDir.mkdir();
-        }
+        if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
 
-        abc:
-        boolean check = new File(homeDir, "yarsaa/"+fileName).exists()
-        if (check == true) {
-            Matcher m = PATTERN.matcher(fileName);
-            if (m.matches()) {
-                String prefix = m.group(1);
-                String last = m.group(2);
-                String suffix = m.group(3);
-                if (suffix == null) suffix = "";
-                int count = last != null ? Integer.parseInt(last) : 0;
-                count++;
-                fileName = prefix + "(" + count + ")" + suffix;
-                continue abc
+            def mp = (MultipartHttpServletRequest) request
+            CommonsMultipartFile file = (CommonsMultipartFile) mp.getFile("coverImageName")
+            def fileName = file.originalFilename
+            def homeDir = new File(System.getProperty("user.home"))
+            File theDir = new File(homeDir, "yarsaa");
+            if (!theDir.exists()) {
+                theDir.mkdir();
             }
-        }
-        File fileDest = new File(homeDir,"yarsaa/${fileName}")
-        file.transferTo(fileDest)
-        return fileName
 
-    }
-    def editCoverImage(String imageNameOld){
-        def mp = (MultipartHttpServletRequest) request
-        CommonsMultipartFile file = (CommonsMultipartFile) mp.getFile("coverImageName")
-        def homeDir = new File(System.getProperty("user.home"))
-        File theDir = new File(homeDir,"yarsaa");
-        if (! theDir.exists()){
-            theDir.mkdir();
-            print"yes"
-        }
-        if(file.size>0){
-            File fileOld= new File(homeDir,"yarsaa/${imageNameOld}")
-            fileOld.delete();
-            String fileName = file.originalFilename
             abc:
-            boolean check = new File(homeDir, "yarsaa/"+fileName).exists()
+            boolean check = new File(homeDir, "yarsaa/" + fileName).exists()
             if (check == true) {
                 Matcher m = PATTERN.matcher(fileName);
                 if (m.matches()) {
@@ -119,31 +108,74 @@ if(productSubCategoryInstance){
                     continue abc
                 }
             }
-            File fileDest = new File(homeDir,"yarsaa/${fileName}")
+            File fileDest = new File(homeDir, "yarsaa/${fileName}")
             file.transferTo(fileDest)
             return fileName
-
-        }
-        else{
-            return imageNameOld
         }
     }
+    def editCoverImage(String imageNameOld){
+        if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+
+            def mp = (MultipartHttpServletRequest) request
+            CommonsMultipartFile file = (CommonsMultipartFile) mp.getFile("coverImageName")
+            def homeDir = new File(System.getProperty("user.home"))
+            File theDir = new File(homeDir, "yarsaa");
+            if (!theDir.exists()) {
+                theDir.mkdir();
+                print "yes"
+            }
+            if (file.size > 0) {
+                File fileOld = new File(homeDir, "yarsaa/${imageNameOld}")
+                fileOld.delete();
+                String fileName = file.originalFilename
+                abc:
+                boolean check = new File(homeDir, "yarsaa/" + fileName).exists()
+                if (check == true) {
+                    Matcher m = PATTERN.matcher(fileName);
+                    if (m.matches()) {
+                        String prefix = m.group(1);
+                        String last = m.group(2);
+                        String suffix = m.group(3);
+                        if (suffix == null) suffix = "";
+                        int count = last != null ? Integer.parseInt(last) : 0;
+                        count++;
+                        fileName = prefix + "(" + count + ")" + suffix;
+                        continue abc
+                    }
+                }
+                File fileDest = new File(homeDir, "yarsaa/${fileName}")
+                file.transferTo(fileDest)
+                return fileName
+
+            } else {
+                return imageNameOld
+            }
+        }    }
     def show(Long id){
         try{
-        def productSubCategoryInstance=ProductSubCategory.get(id)
+            if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+
+                def productSubCategoryInstance=ProductSubCategory.get(id)
 
         if(productSubCategoryInstance){
             [productSubCategoryInstance:productSubCategoryInstance]}
         else{
             redirect(action: "list")
         }}
+            else{
+                redirect(action: "adminLoginForm",controller: "login")
+
+            }
+        }
         catch (Exception e){
             redirect(action: "notfound",controller: "errorPage")
         }
     }
     def edit(){
         try{
-        def productSubCategoryInstance=ProductSubCategory.get(params.id)
+            if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
+
+                def productSubCategoryInstance=ProductSubCategory.get(params.id)
 
         if(productSubCategoryInstance){
             [productSubCategoryInstance:productSubCategoryInstance]
@@ -151,6 +183,11 @@ if(productSubCategoryInstance){
         else{
             redirect(action: "list")
         }}
+            else{
+                redirect(action: "adminLoginForm",controller: "login")
+
+            }
+            }
         catch (Exception e){
             redirect(action: "notfound",controller: "errorPage")
 
@@ -158,28 +195,32 @@ if(productSubCategoryInstance){
     }
     def delete(){
         try{
+            if (session.adminUser.role == "CEO" || session.adminUser.role == "MD" || session.adminUser.role=="Content Manager") {
 
-            def productSubCategoryInstance=ProductSubCategory.get(params.id)
-
-
-        if(productSubCategoryInstance) {
-                productSubCategoryInstance.delete(flush: true)
-            def imageName=productSubCategoryInstance.coverImageName
-            def homeDir = new File(System.getProperty("user.home"))
-            File file= new File(homeDir,"yarsaa/${imageName}")
-            file.delete();
-                flash.message="Successfully deleted."
+                def productSubCategoryInstance = ProductSubCategory.get(params.id)
 
 
-        }
+                if (productSubCategoryInstance) {
+                    productSubCategoryInstance.delete(flush: true)
+                    def imageName = productSubCategoryInstance.coverImageName
+                    def homeDir = new File(System.getProperty("user.home"))
+                    File file = new File(homeDir, "yarsaa/${imageName}")
+                    file.delete();
+                    flash.message = "Successfully deleted."
+
+
+                } else {
+                    flash.message = "Unable to delete the already deleted item."
+
+
+                }
+                redirect(action: "list")
+
+            }
         else{
-            flash.message="Unable to delete the already deleted item."
-
-
+                redirect(action: "adminLoginForm",controller: "login")
+            }
         }
-        redirect(action: "list")
-
-    }
         catch (DataIntegrityViolationException e){
             flash.message="Sorry! cannot delete this data."
             redirect(action: "list")
